+++ gui/cclock.cc
@@ -80,7 +80,7 @@ void CcLock::Enter()
     if (!( m_CSMutex -> Lock() == wxMUTEX_NO_ERROR) )
     {
        ostringstream strm;
-       strm << "----Failed to lock mutex " << (long)this << "\n";
+       strm << "----Failed to lock mutex " << (uintptr_t)this << "\n";
        LOGSTAT (strm.str());
     }
     m_Locked ++;
